change: move Divider component to be generic
