+++ frontend/src/components/Divider.jsx
