+++ frontend/src/pages/Login.jsx
@@ -9,7 +9,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 // Components
 import Card from "../components/Card";
 import FormInput from "../components/forms/FormInput";
-import Divider from "../components/forms/Divider";
+import Divider from "../components/Divider";
 import SubmitButton from "../components/forms/SubmitButton";
 import Button from "../components/forms/Button";
